requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

alt-profanity-check==1.1.2
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
dill==0.3.5.1
distlib==0.3.4
filelock==3.7.0
Flask==2.2.2
googletrans==3.0.0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.23.0
huggingface-hub==0.8.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
lxml==4.9.1
MarkupSafe==2.1.1
multiprocess==0.70.13
nltk==3.7
numpy==1.23.1
outcome==1.1.0
packaging==21.3
pathos==0.2.9
Pillow==9.2.0
platformdirs==2.5.2
playwright==1.22.0
pox==0.3.1
ppft==1.7.6.5
pycodestyle==2.8.0
pycparser==2.21
pyee==8.1.0
PyExecJS==1.5.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
rfc3986==1.5.0
scikit-learn==1.1.2
scipy==1.9.0
selenium==4.2.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tqdm==4.64.0
transformers==4.21.1
translators==5.4.6
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
virtualenv==20.14.1
waitress==2.1.2
websockets==10.1
Werkzeug==2.2.2
wsproto==1.1.0
